TSK-967: Merge module taskana-rest-spring and taskana-rest-spring-test
