+++ lib/pom.xml
@@ -15,8 +15,8 @@
 	</parent>
 
 	<modules>
-		<module>taskana-core</module>
 		<module>taskana-data</module>
+		<module>taskana-core</module>
 		<module>taskana-cdi</module>
 		<module>taskana-spring</module>
 		<module>taskana-cdi-example</module>
