+++ rest/pom.xml
@@ -18,7 +18,6 @@
 		<module>taskana-rest-spring</module>
 		<module>../web</module>
 		<module>taskana-rest-spring-example-common</module>
-		<module>taskana-rest-spring-test</module>
 		<module>taskana-rest-spring-example-boot</module>
 		<module>taskana-rest-spring-example-wildfly</module>
 	</modules>
