+++ rest/taskana-rest-spring-example-common/pom.xml
@@ -87,6 +87,7 @@
 		<dependency>
 			<groupId>com.h2database</groupId>
 			<artifactId>h2</artifactId>
+			<version>${version.h2}</version>
 			<scope>test</scope>
 		</dependency>
