--- rest/taskana-rest-spring-test/src/test/resources/application.properties
@@ -1,41 +1,0 @@
-logging.level.pro.taskana=INFO
-### logging.level.org.springframework=DEBUG
-######## Taskana DB #######
-datasource.url=jdbc:h2:mem:taskana;IGNORECASE=TRUE;LOCK_MODE=0
-datasource.driverClassName=org.h2.Driver
-datasource.username=sa
-datasource.password=sa
-taskana.schemaName=TASKANA
-####### property that control rest api security deploy use true for no security.
-devMode=false
-
-####### Property that informs about the Taskana's version. This version is shown the application web
-version=@project.version@
-
-####### control LDAP usage
-taskana.ldap.useLdap=false
-####### properties to connect to LDAP
-taskana.ldap.serverUrl=ldap://localhost:10389
-taskana.ldap.bindDn=uid=admin,ou=system
-taskana.ldap.bindPassword=secret
-taskana.ldap.baseDn=o=TaskanaTest
-####### properties that control search for users and groups
-taskana.ldap.userSearchBase=ou=people
-taskana.ldap.userSearchFilterName=objectclass
-taskana.ldap.userSearchFilterValue=person
-taskana.ldap.userFirstnameAttribute=givenName
-taskana.ldap.userLastnameAttribute=sn
-taskana.ldap.userIdAttribute=uid
-taskana.ldap.groupSearchBase=ou=groups
-taskana.ldap.groupSearchFilterName=objectclass
-taskana.ldap.groupSearchFilterValue=groupOfUniqueNames
-taskana.ldap.groupNameAttribute=cn
-taskana.ldap.minSearchForLength=3
-taskana.ldap.maxNumberOfReturnedAccessIds=50
-####### JobScheduler cron expression that specifies when the JobSchedler runs
-taskana.jobscheduler.async.cron=0 0 * * * *
-####### cache static resources properties
-spring.resources.cache.cachecontrol.cache-private=true
-####### tomcat is not detecting the x-forward headers from bluemix as a trustworthy proxy
-server.tomcat.internal-proxies=.*
-server.use-forward-headers=true
