+++ rest/taskana-rest-spring/src/test/java/pro/taskana/rest/MappingTest.java
