--- rest/taskana-rest-spring/src/test/java/pro/taskana/rest/resource/ConfigurationTest.java
@@ -1,18 +1,0 @@
-package pro.taskana.rest.resource;
-
-import static org.junit.Assert.assertNotNull;
-
-import org.junit.jupiter.api.Test;
-
-import pro.taskana.configuration.TaskanaEngineConfiguration;
-
-/**
- * Tests for Taskana lib configuration.
- */
-class ConfigurationTest {
-
-    @Test
-    void testImplementationVersionIsInTaskanaCorePackage() {
-        assertNotNull(TaskanaEngineConfiguration.class.getPackage().getImplementationVersion());
-    }
-}
