+++ rest/taskana-rest-spring/src/test/java/pro/taskana/rest/resource/ReportResourceTest.java
@@ -15,25 +15,20 @@ import java.util.stream.IntStream;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.context.web.WebAppConfiguration;
 
+import pro.taskana.TaskanaSpringBootTest;
 import pro.taskana.impl.report.header.TimeIntervalColumnHeader;
 import pro.taskana.impl.report.item.DetailedMonitorQueryItem;
 import pro.taskana.impl.report.item.MonitorQueryItem;
 import pro.taskana.report.ClassificationReport;
 import pro.taskana.report.WorkbasketReport;
-import pro.taskana.rest.TestConfiguration;
 
 /**
  * Test for {@link ReportResourceAssembler}.
  */
-@ExtendWith(SpringExtension.class)
-@ContextConfiguration(classes = {TestConfiguration.class})
-@WebAppConfiguration
+
+@TaskanaSpringBootTest
 class ReportResourceTest {
 
     @Autowired
