+++ rest/taskana-rest-spring/src/test/java/pro/taskana/rest/resource/WorkbasketAccessItemResourceAssemblerTest.java
@@ -2,23 +2,18 @@ package pro.taskana.rest.resource;
 
 import org.junit.Assert;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.context.web.WebAppConfiguration;
 
+import pro.taskana.TaskanaSpringBootTest;
 import pro.taskana.WorkbasketAccessItem;
 import pro.taskana.WorkbasketService;
 import pro.taskana.impl.WorkbasketAccessItemImpl;
-import pro.taskana.rest.TestConfiguration;
 
 /**
  * Test for {@link WorkbasketAccessItemResourceAssembler}.
  */
-@ExtendWith(SpringExtension.class)
-@ContextConfiguration(classes = {TestConfiguration.class})
-@WebAppConfiguration
+
+@TaskanaSpringBootTest
 class WorkbasketAccessItemResourceAssemblerTest {
 
     @Autowired
