+++ rest/taskana-rest-spring/src/test/java/pro/taskana/rest/security/WebSecurityConfig.java
